moved posts ui in a different file
