+++ app/src/main/java/com/example/needlevision/PostActivity.java
@@ -1,0 +1,61 @@
+package com.example.needlevision;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+
+import com.example.needlevision.adapters.SlidePagerAdapter;
+import com.example.needlevision.fragments.map_fragment;
+import com.example.needlevision.fragments.posts_fragment;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class PostActivity extends AppCompatActivity {
+    private ViewPager pager;
+    private PagerAdapter pagerAdapter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_post);
+        loadPagerPage();
+    }
+
+    // Loads the Fragment Slides
+    public void loadPagerPage(){
+        List<Fragment> list = new ArrayList<>();
+        list.add(new map_fragment());
+        list.add(new posts_fragment());
+
+        pager = findViewById(R.id.pager);
+        pagerAdapter = new SlidePagerAdapter(getSupportFragmentManager(), list);
+        pager.setAdapter(pagerAdapter);
+
+        loadFAB();
+    }
+
+    private void loadFAB(){
+        FloatingActionButton fab = findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Bring Up Camera", Snackbar.LENGTH_LONG)
+                        .setAction("Action", null).show();
+
+            }
+        });
+    }
+
+    private void loadImageDetails(){
+//        Intent intent =new Intent(PostActivity.this, );
+
+    }
+}
