+++ app/src/main/res/layout/activity_main.xml
@@ -7,23 +7,4 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-
-    <androidx.viewpager.widget.ViewPager
-        android:id="@+id/pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true"
-        android:layout_margin="16dp"
-        android:contentDescription="@string/camera"
-        android:src="@android:drawable/ic_menu_camera"/>
-
-
-
 </RelativeLayout>
